+++ src/gl_boilerplate.hpp
@@ -36,8 +36,8 @@ void render_loop(
   GLFWwindow *window,
 
   const std::function<void(
-    const GLfloat&,
-    const GLfloat&,
+    const GLint&,
+    const GLint&,
     const double&
   )> render_callback,
